+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 - #60 - Added new provider `Seaborn` using `Html` and a new parser for Email Subject: `EmailSubjectParser`
 - #61 - Added new provider `Colt` using `ICal` and `Csv`
 - #66 - Added new provider `Momentum` using `Html` and `EmailSubjectParser`
+- #73 - Added new provider `Sparkle` using `Html` and `EmailSubjectParser`. Added support for multiple maintenances with `CombinedProcessor`.
 
 ### Fixed